+++ app/src/main/java/org/sopt/santamanitto/MetricsUtil.kt
@@ -5,25 +5,25 @@ import android.content.res.Resources
 import android.util.DisplayMetrics
 
 object MetricsUtil {
-    fun convertDpToPixel(dp: Int, context: Context?): Int {
+    fun convertDpToPixel(dp: Float, context: Context?): Float {
         return if (context != null) {
             val resources = context.resources
             val metrics = resources.displayMetrics
-            (dp * (metrics.densityDpi.toFloat() / DisplayMetrics.DENSITY_DEFAULT)).toInt()
+            (dp * (metrics.densityDpi.toFloat() / DisplayMetrics.DENSITY_DEFAULT))
         } else {
             val metrics = Resources.getSystem().displayMetrics
-            (dp * (metrics.densityDpi.toFloat() / DisplayMetrics.DENSITY_DEFAULT)).toInt()
+            (dp * (metrics.densityDpi.toFloat() / DisplayMetrics.DENSITY_DEFAULT))
         }
     }
 
-    fun convertPixelsToDp(px: Int, context: Context?): Int {
+    fun convertPixelsToDp(px: Float, context: Context?): Float {
         return if (context != null) {
             val resources = context.resources
             val metrics = resources.displayMetrics
-            (px / (metrics.densityDpi.toFloat() / DisplayMetrics.DENSITY_DEFAULT)).toInt()
+            (px / (metrics.densityDpi.toFloat() / DisplayMetrics.DENSITY_DEFAULT))
         } else {
             val metrics = Resources.getSystem().displayMetrics
-            (px / (metrics.densityDpi.toFloat() / DisplayMetrics.DENSITY_DEFAULT)).toInt()
+            (px / (metrics.densityDpi.toFloat() / DisplayMetrics.DENSITY_DEFAULT))
         }
     }
 }
